SPA: refactor angular application
