--- cap/static/js/cap/cap.records.js
@@ -1,12 +1,0 @@
-(function (angular) {
-
-  // Setup everyhting
-  angular.module('capRecords', [
-  'schemaForm',
-  'invenioRecords.factories',
-  'invenioRecords.services',
-  'invenioRecords.controllers',
-  'invenioRecords.directives',
-  ]);
-
-})(angular);
