--- cap/static/js/cap/cap.search.js
@@ -1,10 +1,0 @@
-(function (angular) {
-
-  // Setup everyhting
-  angular.module('capSearch', [
-    'invenioSearch.services',
-    'invenioSearch.controllers',
-    'invenioSearch.directives'
-  ]);
-
-})(angular);
