+++ cap/templates/cap/page.html
@@ -27,7 +27,6 @@
 {%- extends "cap/page_base.html" %}
 
 {%- block css %}
-{% assets "cap_theme_css" %}<link href="{{ ASSET_URL }}" rel="stylesheet">{% endassets %}
 {% assets "invenio_search_ui_search_css" %}<link href="{{ ASSET_URL }}" rel="stylesheet">{% endassets %}
 <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
 <!--[if lt IE 9]>
